add account and category services

+++ family-money-model/src/main/java/com/familymoney/model/BotUser.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class BotUser {
     private Long id;
-    private Integer telegramId;
+    private Long telegramId;
     private String userName;
     private String firstName;
     private String lastName;
